components/Button: Pull out ripple effect into own utility
so it can be used in other components
